add basic data-export functionality
Separates the downsampling/data-export logic from the batching logic and sets up the basic framework for each "processor" (ds/export) to (a) run concurrently, and (b) share/re-use data that might be expensive to page/compute.
Still very much a WIP.
