+++ core/src/main/resources/filodb-defaults.conf
@@ -420,6 +420,25 @@ filodb {
       #  tagB3 = value3
       #}
     ]
+
+    # TODO(a_theimer): comments
+    # TODO(a_theimer): make types more appropriate
+    # TODO(a_theimer): call "export-specs"
+    export-filters = [
+      [
+        [
+          "label1", "foo",
+          "label2", "bar"
+        ], [
+          "foo1.bar1.com",
+          "foo2.bar2.com"
+        ]
+      ]
+    ]
+    export-struct = [
+      "{{_ws_}}-suffix1"
+      "{{_ns_}}-suffix2"
+    ]
   }
 
   ds-index-job {
